+++ src/restfw/views.py
@@ -14,12 +14,12 @@ from pyramid.traversal import quote_path_segment
 from zope.interface import implementer
 
 from . import interfaces, schemas
-from .errors import ParameterError
+from .errors import MultiParametersError, ParameterError
 from .external_links import get_external_links
 from .hal import HalResource, SimpleContainer
 from .resources import Resource
 from .typing import Json, PyramidRequest
-from .utils import create_validation_error, get_input_data, get_paging_links
+from .utils import create_multi_validation_error, create_validation_error, get_input_data, get_paging_links
 
 
 def get_resource_view(resource, request: PyramidRequest) -> Optional['ResourceView']:
@@ -147,6 +147,11 @@ class ResourceView:
         params = self._get_params()
         try:
             result, created = self.resource.http_post(self.request, params)
+        except MultiParametersError as e:
+            raise create_multi_validation_error(
+                self.options_for_post.input_schema,
+                errors=e.errors,
+            ) from e
         except ParameterError as e:
             raise create_validation_error(
                 self.options_for_post.input_schema,
@@ -162,6 +167,11 @@ class ResourceView:
         params = self._get_params()
         try:
             created = self.resource.http_put(self.request, params)
+        except MultiParametersError as e:
+            raise create_multi_validation_error(
+                self.options_for_put.input_schema,
+                errors=e.errors,
+            ) from e
         except ParameterError as e:
             raise create_validation_error(
                 self.options_for_put.input_schema,
@@ -178,6 +188,11 @@ class ResourceView:
         params = self._get_params()
         try:
             created = self.resource.http_patch(self.request, params)
+        except MultiParametersError as e:
+            raise create_multi_validation_error(
+                self.options_for_patch.input_schema,
+                errors=e.errors,
+            ) from e
         except ParameterError as e:
             raise create_validation_error(
                 self.options_for_patch.input_schema,
@@ -194,6 +209,11 @@ class ResourceView:
         params = self._get_params()
         try:
             result = self.resource.http_delete(self.request, params)
+        except MultiParametersError as e:
+            raise create_multi_validation_error(
+                self.options_for_delete.input_schema,
+                errors=e.errors,
+            ) from e
         except ParameterError as e:
             raise create_validation_error(
                 self.options_for_delete.input_schema,